pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-dance>=7.1.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "oauthlib>=3.3.1",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "flask-login>=0.6.3",
    "werkzeug>=3.1.3",
    "jinja2>=3.1.6",
    "sqlalchemy>=2.0.41",
    "requests>=2.32.4",
    "urllib3>=2.5.0",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.1",
    "mysql-connector-python>=9.3.0",
    "pymysql>=1.1.1",
    "qrcode[pil]>=8.2",
    "pillow>=11.3.0",
]
